refactor: theme에서 standardValues분리
