+++ src/styles/index.jsx
@@ -1,2 +1,3 @@
 export { default as mixin } from './mixin';
 export { default as theme } from './theme';
+export { default as standardValues } from './standardValues';
