+++ src/styles/standardValues.jsx
@@ -1,0 +1,26 @@
+const standardValues = {
+  card: {
+    maxWidth: '330px',
+    margin: '20px',
+    maxCount: {
+      top: 4,
+      tab: 3,
+      mobL: 2,
+      mobS: 1,
+    },
+  },
+  userList: {
+    maxCount: 10,
+    size: '5rem',
+  },
+  banner: {
+    maxCount: {
+      top: 5,
+      tab: 3,
+      mobL: 2,
+      mobS: 1,
+    },
+  },
+};
+
+export default standardValues;
